scripts/requirements.txt
feedparser>=6.0
requests>=2.28
arxiv>=2.0
python-dateutil
markdown
deep-translator>=3.4
